Move filtering of automatic headers from HttpClient to PerRequest
